뷰 리팩토링
