+++ build.gradle
@@ -24,6 +24,8 @@ dependencies {
     implementation 'org.springframework.boot:spring-boot-starter-web'
     implementation 'org.springframework.boot:spring-boot-starter-validation'
 
+    developmentOnly 'org.springframework.boot:spring-boot-devtools'
+
     implementation('com.github.gavlyukovskiy:p6spy-spring-boot-starter:1.6.2')
 
     compileOnly 'org.projectlombok:lombok'
