+++ src/main/resources/application-dev.properties
@@ -1,0 +1,8 @@
+#절대경로
+project.base-dir=file:////Users/yuhyeonho/Desktop/code/cafe-member-calcWage
+#정적파일(javascript.css등등) 위치
+spring.web.resources.static-locations=${project.base-dir}/src/main/resources/static/
+#템플릿HTML이 포함된 URL
+spring.thymeleaf.prefix=${project.base-dir}/src/main/resources/templates/
+#실제로 설정해주는 부분
+spring.thymeleaf.cache=false
