+++ src/main/resources/application.yml
@@ -1,10 +1,11 @@
 spring:
-  datasource:
-    url: jdbc:h2:tcp://localhost/~/cafe
-    username: sa
-    password:
-    driver-class-name: org.h2.Driver
-
+#  datasource:
+#    url: jdbc:h2:tcp://localhost/~/cafe
+#    username: sa
+#    password:
+#    driver-class-name: org.h2.Driver
+  profiles:
+    include: dev
   jpa:
     hibernate:
       ddl-auto: update
@@ -19,3 +20,4 @@ logging:
   level:
     org.hibernate.SQL: debug
     org.hibernate.type: trace
+
