--- src/main/resources/static/css/createMemberForm.css
@@ -1,38 +1,0 @@
-.fieldError {
-    border-color: #bd2130;
-    color: red;
-}
-
-.container {
-    margin-top: 50px;
-    background-color: gray;
-}
-.input-box {
-    width: 70%;
-    margin-left: 48px;
-    text-align: center;
-}
-.head-text {
-    margin-left: 5px;
-    margin-top: 5px;
-    margin-bottom: 20px;
-    text-align: left;
-    font-size: 24pt;
-    letter-spacing: 5px;
-}
-
-.btn {
-    /*position: absolute;!**!*/
-    left: 10px;
-    margin-top: 30px;
-    top: 10px;
-    padding: 10px;
-    font-size: 16pt;
-    letter-spacing: 7px;
-    background-color: #3e70ff;
-    border: none;
-    color: white;
-    border-radius: 5px;
-    width: 100%;
-    align-content: center;
-}
