+++ src/main/resources/static/css/members/form.css
@@ -1,0 +1,60 @@
+body {
+  font-family: 'Poppins', sans-serif;
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  min-height: 100px;
+  font-size: 1.5rem;
+  background-color: #222222;
+}
+
+input[type=text] {
+  width: 100%;
+  padding: 12px 20px;
+  margin: 8px 0;
+  display: inline-block;
+  border: 1px solid #ccc;
+  border-radius: 4px;
+  box-sizing: border-box;
+  font-size: 15pt;
+}
+
+.fieldError {
+  border-color: #bd2130;
+  color: red;
+}
+
+.container {
+  margin-top: 50px;
+  background-color: gray;
+  width: 350px;
+}
+.input-box {
+  width: 70%;
+  margin-left: 48px;
+  text-align: center;
+}
+.head-text {
+  margin-left: 5px;
+  margin-top: 5px;
+  margin-bottom: 20px;
+  text-align: center;
+  font-size: 24pt;
+  letter-spacing: 5px;
+}
+
+.btn {
+  /*position: absolute;!**!*/
+  left: 10px;
+  margin-top: 30px;
+  top: 10px;
+  padding: 10px;
+  font-size: 16pt;
+  letter-spacing: 7px;
+  background-color: #3e70ff;
+  border: none;
+  color: white;
+  border-radius: 5px;
+  width: 100%;
+  align-content: center;
+}
