+++ src/main/resources/static/css/members/info.css
@@ -20,3 +20,13 @@ input {
     Top:0;
     Left :0
 }
+
+body {
+    font-family: 'Poppins', sans-serif;
+    display: flex;
+    flex-direction: column;
+    align-items: center;
+    min-height: 100vh;
+    font-size: 1.5rem;
+    background-color:#222222;
+}
