+++ src/main/resources/static/css/members/total.css
@@ -1,0 +1,16 @@
+body {
+  font-family: 'Poppins', sans-serif;
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  min-height: 100vh;
+  font-size: 1.5rem;
+  background-color:#222222;
+}
+td, th {
+  padding: 10px 10px 10px 10px;
+  color: white;
+}
+.tb {
+  margin-top: 30px;
+}
