+++ src/main/resources/templates/fragments.html
@@ -1,0 +1,14 @@
+<!DOCTYPE html>
+<html lang="en" xmlns:th="http://www.thymeleaf.org">
+<head>
+    <meta charset="UTF-8">
+    <title>Title</title>
+</head>
+<body>
+
+<div th:fragment="logo">
+    <a class="title" th:href="@{/}">COFFEE HOLE</a>
+</div>
+
+</body>
+</html>
