--- src/main/resources/templates/member/createMemberForm.html
@@ -1,106 +1,0 @@
-<!DOCTYPE html>
-<html lang="en" xmlns:th="http://www.thymeleaf.org">
-<!--<link rel="stylesheet" href="/css/createMemberForm.css"/>-->
-<link rel="stylesheet" href="/css/head.css"/>
-<style>
-    body {
-        font-family: 'Poppins', sans-serif;
-        display: flex;
-        flex-direction: column;
-        align-items: center;
-        min-height: 100px;
-        font-size: 1.5rem;
-        background-color: #222222;
-    }
-
-    input[type=text] {
-        width: 100%;
-        padding: 12px 20px;
-        margin: 8px 0;
-        display: inline-block;
-        border: 1px solid #ccc;
-        border-radius: 4px;
-        box-sizing: border-box;
-    }
-
-    .fieldError {
-        border-color: #bd2130;
-        color: red;
-    }
-
-    .container {
-        margin-top: 50px;
-        background-color: gray;
-    }
-    .input-box {
-        width: 70%;
-        margin-left: 48px;
-        text-align: center;
-        font-size: 15pt;
-    }
-    .head-text {
-        margin-left: 5px;
-        margin-top: 5px;
-        margin-bottom: 20px;
-        text-align: left;
-        font-size: 24pt;
-        letter-spacing: 5px;
-    }
-
-    .btn {
-        /*position: absolute;!**!*/
-        left: 10px;
-        margin-top: 30px;
-        top: 10px;
-        padding: 10px;
-        font-size: 16pt;
-        letter-spacing: 7px;
-        background-color: #3e70ff;
-        border: none;
-        color: white;
-        border-radius: 5px;
-        width: 100%;
-        align-content: center;
-    }
-</style>
-<head>
-    <meta charset="UTF-8">
-    <title>Title</title>
-</head>
-
-<body>
-<div>
-    <div>
-        <button class="title" onclick="location.href='/'">COFFEE HOLE</button>
-    </div>
-
-    <!--    <h2>직원 등록</h2>-->
-
-    <form action="/member/new" method="post" th:object="${memberForm}">
-        <div class="container">
-            <div>
-                <div class="head-text">
-                    직원등록
-                </div>
-
-                <div class="input-box">
-                    <input type="text" id="name" placeholder="직원의 이름을 입력해주세요"
-                           th:field="*{name}"
-                           th:class="${#fields.hasErrors('name')}?'form-control fieldError' : 'form-control'"/>
-                    <label class="fieldError" th:for="name"
-                           th:if="${#fields.hasErrors('name')}"
-                           th:errors="*{name}"> Incorrect date
-                    </label>
-                </div>
-                <div>
-                    <button class="btn" type="submit">추가</button>
-                </div>
-            </div>
-
-        </div>
-    </form>
-
-</div>
-</body>
-</html>
-
