--- src/main/resources/templates/member/listForm.html
@@ -1,66 +1,0 @@
-<!DOCTYPE html>
-<html lang="en" xmlns:th="http://www.thymeleaf.org">
-<link rel="stylesheet" href="/css/head.css"/>
-
-<head>
-    <meta charset="UTF-8">
-    <title>Title</title>
-    <style>
-        thead {
-            background: #c8c8c8;
-        }
-
-        th {
-            padding: 8px 8px 8px 8px;
-            background : #2d60f3;
-            font-weight: bold;
-            color: white;
-        }
-        td {
-            padding: 10px 10px 10px 10px;
-            color: white;
-        }
-
-        body {
-            font-family: 'Poppins', sans-serif;
-            display: flex;
-            flex-direction: column;
-            align-items: center;
-            min-height: 100vh;
-            font-size: 1.5rem;
-            background-color:#222222;
-        }
-    </style>
-</head>
-<body>
-
-<div style="text-align: center;">
-    <div>
-        <button class="title" onclick="location.href='/'">COFFEE HOLE</button>
-    </div>
-
-    <table border="1" style="display: inline-block; border-collapse: collapse; margin: 5px;">
-        <thead>
-        <tr>
-            <th>#</th>
-            <th>이름</th>
-            <th>수정</th>
-            <th>삭제</th>
-        </tr>
-        </thead>
-        <tbody>
-        <tr th:each="member : ${members}">
-            <td th:text="${memberStat.count}"></td>
-            <td th:text="${member.name}"></td>
-            <td>
-                <a href="#" th:href="@{/members/member/{id}/update (id=${member.id})}">수정</a>
-            </td>
-            <td>
-                <a href="#" th:href="@{/members/member/{id}/delete (id=${member.id})}">삭제</a>
-            </td>
-        </tr>
-        </tbody>
-    </table>
-</div>
-</body>
-</html>
