+++ src/main/resources/templates/members/total.html
@@ -1,36 +1,14 @@
 <!DOCTYPE html>
 <html lang="en" xmlns:th="http://www.thymeleaf.org">
-<link rel="stylesheet" href="/css/members.css">
-<link rel="stylesheet" href="/css/head.css">
-<style>
-    body {
-        font-family: 'Poppins', sans-serif;
-        display: flex;
-        flex-direction: column;
-        align-items: center;
-        min-height: 100vh;
-        font-size: 1.5rem;
-        background-color:#222222;
-    }
-    td {
-        padding: 10px 10px 10px 10px;
-        color: white;
-    }
-    .tb {
-        margin-top: 30px;
-    }
-</style>
 <head>
     <meta charset="UTF-8">
     <title>Title</title>
+    <link rel="stylesheet" href="/css/head.css">
+    <link rel="stylesheet" href="/css/members/total.css">
 </head>
-
 <body>
+<div th:replace="fragments :: logo"></div>
 <div style="text-align: center;">
-    <div>
-        <button class="title" onclick="location.href='/'">COFFEE HOLE</button>
-    </div>
-
     <table class="tb" border="1">
         <th>번호</th>
         <th>이름</th>
